+++ admin.php
@@ -6,6 +6,7 @@ include LEAFEXT_PLUGIN_DIR . '/admin/elevation/main.php';
 include LEAFEXT_PLUGIN_DIR . '/admin/cluster/main.php';
 include LEAFEXT_PLUGIN_DIR . '/admin/gesture.php';
 include LEAFEXT_PLUGIN_DIR . '/admin/tiles/main.php';
+include LEAFEXT_PLUGIN_DIR . '/admin/canvas.php';
 
 // Add menu page
 function leafext_add_page() {
@@ -96,6 +97,7 @@ function leafext_do_page() {
 		leafext_admin_tiles($active_tab);
 	} else if( $active_tab == 'hover' ) {
 		include LEAFEXT_PLUGIN_DIR . '/admin/help/hovergeojson.php';
+		leafext_canvas_do_page ();
 	} else if( $active_tab == 'gesture' ) {
 		echo '<form method="post" action="options.php">';
 		settings_fields('leafext_settings_gesture');