+++ php/hovergeojson.php
@@ -242,10 +242,25 @@ window.WPLeafletMapPlugin.push(function () {
 	return "\n".$text."\n";
 }
 
+function leafext_canvas_script($tolerance) {
+	return '<script>
+	  window.WPLeafletMapPlugin = window.WPLeafletMapPlugin || [];
+	  window.WPLeafletMapPlugin.push(function () {
+	    var map = window.WPLeafletMapPlugin.getCurrentMap();
+	    map.options.renderer=L.canvas({ tolerance: '.$tolerance.' });
+	  });
+	</script>';
+}
+
 function leafext_geojsonhover_function($atts){
-	$exclude = shortcode_atts( array('exclude' => false), $atts);
-	$text=leafext_geojsonhover_script($exclude['exclude']);
+	$settings = shortcode_atts(	array('exclude' => false,'tolerance' => 0), get_option( 'leafext_canvas' ));
+	$options  = shortcode_atts( $settings, $atts);
+	//var_dump($atts,get_option( 'leafext_canvas'),$settings,$options); wp_die();
+	$text = "";
+	if ($options['tolerance'] != 0) {
+		$text = $text.leafext_canvas_script( $options['tolerance'] );
+	}
+	$text=$text.leafext_geojsonhover_script($options['exclude']);
 	return $text;
 }
 add_shortcode('hover', 'leafext_geojsonhover_function' );
-?>